docs: update the README
Update the README giving more prominency to the configuration with
comments on PRs. Also use the major version number of the action.

This should promote more up to date publiccode.yml files because
deprecations and other warning messages will be shown with the projects'
first PR after a new version of the action.

Also, update old URLs.